+++ R/zzz.R
@@ -5,8 +5,8 @@
     styler.addins_style_transformer = "styler::tidyverse_style()",
     styler.cache_name = styler_version,
     styler.colored_print.vertical = TRUE,
-    styler.ignore_start = c("# styler: off", "# nolint", "nolint start"),
-    styler.ignore_stop = c("# styler: on", "# nolint end"),
+    styler.ignore_start = "styler: off",
+    styler.ignore_stop = "styler: on",
     styler.quiet = FALSE,
     styler.test_dir_writable = TRUE
   )